Simple[Decoder|Encoder|...] -> [Decoder|Encoder|...]
And the file operations will move into "*Plus", e.g. DecoderPlus

I'd rather the core functionality be in the class with the more obvious
name
