+++ src/lib/encoder/CMakeLists.txt
@@ -2,10 +2,10 @@ cmake_minimum_required(VERSION 3.10)
 
 set(SOURCES
 	Decoder.h
+	DecoderPlus.h
 	Encoder.h
+	EncoderPlus.h
 	ReedSolomon.h
-	SimpleDecoder.h
-	SimpleEncoder.h
 	aligned_stream.h
 	escrow_buffer_writer.h
 	reed_solomon_stream.h
