+++ src/lib/encoder/test/EncoderTest.cpp
@@ -2,7 +2,7 @@
 #include "unittest.h"
 #include "TestHelpers.h"
 
-#include "encoder/Encoder.h"
+#include "encoder/EncoderPlus.h"
 #include "fountain/FountainInit.h"
 #include "image_hash/average_hash.h"
 #include "serialize/format.h"
@@ -23,7 +23,7 @@ TEST_CASE( "EncoderTest/testVanilla", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.vanilla";
 
-	Encoder enc(40, 4, 2);
+	EncoderPlus enc(40, 4, 2);
 	assertEquals( 3, enc.encode(inputFile, outPrefix) );
 
 	std::vector<uint64_t> hashes = {0xe727a520684bccec, 0x46a06f002dcded87, 0x4eb1e3646fce8c08};
@@ -45,7 +45,7 @@ TEST_CASE( "EncoderTest/testFountain.4c", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	Encoder enc(40, 4, 2);
+	EncoderPlus enc(40, 4, 2);
 	enc.set_legacy_mode();
 	assertEquals( 3, enc.encode_fountain(inputFile, outPrefix, 0) );
 
@@ -68,7 +68,7 @@ TEST_CASE( "EncoderTest/testFountain.B", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	Encoder enc(40, 4, 2);
+	EncoderPlus enc(40, 4, 2);
 	assertEquals( 4, enc.encode_fountain(inputFile, outPrefix, 0) );
 
 	std::vector<uint64_t> hashes = {0xcf09eb067c876ea6, 0x4697a76025a40c43, 0x666aaca0ec8d6d43, 0xe6e44ca8ec33a260};
@@ -90,7 +90,7 @@ TEST_CASE( "EncoderTest/testFountain.Compress", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	Encoder enc(30, 4, 2);
+	EncoderPlus enc(30, 4, 2);
 	assertEquals( 1, enc.encode_fountain(inputFile, outPrefix) );
 
 	uint64_t hash = 0xa66a666543280e8e;
@@ -103,7 +103,7 @@ TEST_CASE( "EncoderTest/testPiecemealFountainEncoder", "[unit]" )
 {
 	// use the fountain_encoder_stream directly on a char*,int pair
 	MakeTempDirectory tempdir;
-	Encoder enc(40, 4, 2);
+	EncoderPlus enc(40, 4, 2);
 
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string contents = File(inputFile).read_all();
@@ -130,7 +130,7 @@ TEST_CASE( "EncoderTest/testFountain.Size", "[unit]" )
 	std::string inputFile = TestCimbar::getProjectDir() + "/LICENSE";
 	std::string outPrefix = tempdir.path() / "encoder.fountain";
 
-	Encoder enc(30, 4, 2);
+	EncoderPlus enc(30, 4, 2);
 	assertEquals( 1, enc.encode_fountain(inputFile, outPrefix, 16, 1.6) );
 
 	uint64_t hash = 0xa66a666543280e8e;
