+++ src/lib/extractor/Extractor.cpp
@@ -1,13 +1,13 @@
 /* This code is subject to the terms of the Mozilla Public License, v.2.0. http://mozilla.org/MPL/2.0/. */
-#include "SimpleExtractor.h"
+#include "Extractor.h"
 
-#include "SimpleDeskewer.h"
+#include "Deskewer.h"
 #include "Scanner.h"
 #include "cimb_translator/Config.h"
 #include <vector>
 using std::string;
 
-SimpleExtractor::SimpleExtractor(cimbar::vec_xy image_size, unsigned anchor_size)
+Extractor::Extractor(cimbar::vec_xy image_size, unsigned anchor_size)
 	: _imageSize({
 		image_size.width()? image_size.width() : cimbar::Config::image_size_x(),
 		image_size.height()? image_size.height() : cimbar::Config::image_size_y()})
@@ -15,7 +15,7 @@ SimpleExtractor::SimpleExtractor(cimbar::vec_xy image_size, unsigned anchor_size
 {
 }
 
-int SimpleExtractor::extract(const cv::Mat& img, cv::Mat& out)
+int Extractor::extract(const cv::Mat& img, cv::Mat& out)
 {
 	Scanner scanner(img);
 	std::vector<Anchor> points = scanner.scan();
@@ -23,7 +23,7 @@ int SimpleExtractor::extract(const cv::Mat& img, cv::Mat& out)
 		return FAILURE;
 
 	Corners corners(points);
-	SimpleDeskewer de(_imageSize, _anchorSize);
+	Deskewer de(_imageSize, _anchorSize);
 	out = de.deskew(img, corners);
 
 	if ( !corners.is_granular_scale(_imageSize) )
@@ -31,7 +31,7 @@ int SimpleExtractor::extract(const cv::Mat& img, cv::Mat& out)
 	return SUCCESS;
 }
 
-int SimpleExtractor::extract(const cv::UMat& img, cv::UMat& out)
+int Extractor::extract(const cv::UMat& img, cv::UMat& out)
 {
 	Scanner scanner(img);
 	std::vector<Anchor> points = scanner.scan();
@@ -39,7 +39,7 @@ int SimpleExtractor::extract(const cv::UMat& img, cv::UMat& out)
 		return FAILURE;
 
 	Corners corners(points);
-	SimpleDeskewer de(_imageSize, _anchorSize);
+	Deskewer de(_imageSize, _anchorSize);
 	out = de.deskew(img, corners);
 
 	if ( !corners.is_granular_scale(_imageSize) )
