+++ src/lib/extractor/test/ExtractorTest.cpp
@@ -2,7 +2,7 @@
 #include "unittest.h"
 #include "TestHelpers.h"
 
-#include "Extractor.h"
+#include "ExtractorPlus.h"
 #include "image_hash/average_hash.h"
 #include "util/MakeTempDirectory.h"
 #include <iostream>
@@ -14,7 +14,7 @@ TEST_CASE( "ExtractorTest/testExtract", "[unit]" )
 	MakeTempDirectory tempdir;
 
 	std::string imgPath = tempdir.path() / "ex.jpg";
-	Extractor ext({1024, 1024}, 30);
+	ExtractorPlus ext({1024, 1024}, 30);
 	ext.extract(TestCimbar::getSample("6bit/4_30_f0_big.jpg"), imgPath);
 
 	cv::Mat out = cv::imread(imgPath);
@@ -27,7 +27,7 @@ TEST_CASE( "ExtractorTest/testExtractMid", "[unit]" )
 	MakeTempDirectory tempdir;
 
 	std::string imgPath = tempdir.path() / "ex.jpg";
-	Extractor ext({1024, 1024}, 30);
+	ExtractorPlus ext({1024, 1024}, 30);
 	ext.extract(TestCimbar::getSample("6bit/4_30_f2_734.jpg"), imgPath);
 
 	cv::Mat out = cv::imread(imgPath);
@@ -40,7 +40,7 @@ TEST_CASE( "ExtractorTest/testExtractUpscale", "[unit]" )
 	MakeTempDirectory tempdir;
 
 	std::string imgPath = tempdir.path() / "exup.jpg";
-	Extractor ext({1024, 1024}, 30);
+	ExtractorPlus ext({1024, 1024}, 30);
 	ext.extract(TestCimbar::getSample("6bit/4_30_f0_627.jpg"), imgPath);
 
 	cv::Mat out = cv::imread(imgPath);
